requirements.txt
streamlit==1.42.0
pandas==2.2.3
google-cloud-bigquery==3.29.0
db-dtypes==1.3.1
gcsfs
openpyxl
